+++ Packages/com.espale.ui/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2024-12-25
+
+### Changed
+
+- `UILayer`'s will now auto-disable when a scene starts, if it's not the default active layer.
+- Changed the `canvasGroup` and `canvas` variables from `private` to `protected` to make implementing custom UI Layers easier.
+
 ## [1.0.2] - 2024-12-20
 
 ### Added